Fixed SearchProjectionOnPial.m
The version of SearchProjectionOnPial.m that comes with FreeSurfer
doesn’t work with the recon-all localGI flag. I believe the issue is
that with some versions of free surfer a variable is a row vector but
with others it is a column vector. My slight tweak can deal with either.